+++ R/doConcrete.R
@@ -65,7 +65,8 @@ doConcrete <- function(ConcreteArgs) {
 doConCRTmle <- function(DataTable, TargetTime, TargetEvent, Regime, CVFolds, Model, MaxUpdateIter,
                         OneStepEps, MinNuisance, Verbose, GComp, ReturnModels,
                         UpdateMethod = "standard",
-                        EICStopRule = "relative", EICStopAbsTol = 0, CrossFit = FALSE, ...)
+                        EICStopRule = "relative", EICStopAbsTol = 0, CrossFit = FALSE,
+                        HazEnsemble = FALSE, ...)
 {
     ratio <- Time <- Event <- PnEIC <- `seEIC/(sqrt(n)log(n))` <- NULL # for data.table compatibility w/ global var binding check
     EICStopRule <- getEICStopRule(EICStopRule)
@@ -76,7 +77,7 @@ doConCRTmle <- function(DataTable, TargetTime, TargetEvent, Regime, CVFolds, Mod
     Estimates <- InitFn(Data = DataTable, Model = Model, CVFolds = CVFolds,
                         MinNuisance = MinNuisance, TargetEvent = TargetEvent,
                         TargetTime = TargetTime, Regime = Regime,
-                        ReturnModels = ReturnModels)
+                        ReturnModels = ReturnModels, HazEnsemble = HazEnsemble)
 
     # get initial EIC (possibly with GComp plug-in estimate) ---------------------------------------------
     Estimates <- getEIC(Estimates = Estimates, Data = DataTable, Regime = Regime,